Remove sync method Post from the ZazClient
This method used the underlining PostScheduled method call that caused
some issues in the high load schenarious when lots of concurrent clients
are starting async tasks that sending 3 req/second just to check the task
status. I've closed this method with appropriate comments just to enforce
livrary clients to review all their code that use this method and to
upgrade it to PostAsync use. In case if this kind of upgrade is not
possible they will be able to access the hidden method in through
inheritance.